refactor: simplify.
