+++ pkg/cli/commands_test.go
@@ -679,27 +679,3 @@ Flags:
 
 `, string(out))
 }
-
-type Yo struct {
-	Foo string `description:"Foo description"`
-	Fii string `description:"Fii description"`
-	Fuu string `description:"Fuu description"`
-	Yi  *Yi    `label:"allowEmpty"`
-	Yu  *Yi
-}
-
-func (y *Yo) SetDefaults() {
-	y.Foo = "foo"
-	y.Fii = "fii"
-}
-
-type Yi struct {
-	Foo string
-	Fii string
-	Fuu string
-}
-
-func (y *Yi) SetDefaults() {
-	y.Foo = "foo"
-	y.Fii = "fii"
-}
