+++ pkg/cli/loader_flag.go
@@ -11,7 +11,7 @@ import (
 type FlagLoader struct{}
 
 // Load loads the configuration.
-func (f *FlagLoader) Load(args []string, cmd *Command) (bool, error) {
+func (*FlagLoader) Load(args []string, cmd *Command) (bool, error) {
 	if err := flag.Decode(args, cmd.Configuration); err != nil {
 		return false, fmt.Errorf("failed to decode configuration from flags: %v", err)
 	}
